requirements.txt
psycopg2-binary~=2.9.10
python-dotenv~=1.0.1
SQLAlchemy~=2.0.36